Extract doc-build pipeline into reusable package
The logic that transforms a k6-docs checkout into a doc bundle
(sections.json plus a markdown tree) lived inside the cmd/prepare
command's main package, so nothing else could reuse it.

Move it into an importable package, leaving cmd/prepare with just its
flag parsing and the optional repo clone. This is a pure refactor: the
prepare tool produces identical output, and it unlocks building bundles
in-process for upcoming features.
